# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: SMS Blossom API
  version: 1.0.0
  description: >
    REST API for Shopify SMS marketing app (Express, ESM).
    Includes OAuth-driven backend, consent flows, discounts, reporting, webhooks, and GDPR.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.sms-blossom.com
    description: Production API
...
security: []
paths:
...
  /automations:
    get:
      tags: [Automations]
      summary: Read automations config
      operationId: getAutomations
      description: |
        Includes abandoned checkout settings:
        - `abandoned.delayMinutes`: inactivity window before SMS is sent (default 30)
        - `abandoned.discountCode`: optional code appended to recovery URL
      parameters:
        - in: query
          name: shop
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Automations config
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AutomationsResponse' }
        '400':
          description: Missing shop parameter
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Shop not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    put:
      tags: [Automations]
      summary: Update automations config
      operationId: updateAutomations
      parameters:
        - in: query
          name: shop
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AutomationsUpdateRequest' }
      responses:
        '200':
          description: Updated automations config
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AutomationsResponse' }
        '400':
          description: Missing shop parameter
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Shop not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '422':
          description: Invalid payload
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /reports/automations:
    get:
      tags: [Reports]
      summary: Automation attribution (Abandoned Checkout revenue)
      operationId: getAutomationAttribution
      parameters:
        - in: query
          name: shop
          required: true
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: window
          schema: { type: string, example: '30d' }
      responses:
        '200':
          description: Automation attribution
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  range:
                    type: object
                    properties:
                      from: { type: string, format: date-time }
                      to: { type: string, format: date-time }
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        automation: { type: string }
                        orders: { type: integer }
                        revenue: { type: number }
        '404':
          description: Unknown shop
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
components:
  schemas:
...
    AutomationRules:
      type: object
      properties:
        quietHours:
          type: object
          properties:
            enabled: { type: boolean }
            start: { type: integer, minimum: 0, maximum: 23 }
            end: { type: integer, minimum: 0, maximum: 23 }
            zone: { type: [string, 'null'] }
        frequencyCap:
          type: object
          properties:
            enabled: { type: boolean }
            per: { type: string, enum: [hour, day, week] }
            max: { type: integer, minimum: 1 }
        dedupeWindowMin: { type: integer, minimum: 0 }
    AutomationConfig:
      type: object
      properties:
        enabled: { type: boolean }
        template: { type: [string, 'null'] }
        delayMinutes: { type: integer, minimum: 5, maximum: 1440 }
        rules: { $ref: '#/components/schemas/AutomationRules' }
    AutomationsResponse:
      type: object
      properties:
        ok: { type: boolean }
        shop: { type: string }
        automations:
          type: object
          properties:
            orderPaid: { $ref: '#/components/schemas/AutomationConfig' }
            abandoned: { $ref: '#/components/schemas/AutomationConfig' }
            fulfillmentUpdate: { $ref: '#/components/schemas/AutomationConfig' }
            welcome: { $ref: '#/components/schemas/AutomationConfig' }
            backInStock: { $ref: '#/components/schemas/AutomationConfig' }
    AutomationsUpdateRequest:
      type: object
      properties:
        orderPaid: { $ref: '#/components/schemas/AutomationConfig' }
        abandoned: { $ref: '#/components/schemas/AutomationConfig' }
        fulfillmentUpdate: { $ref: '#/components/schemas/AutomationConfig' }
        welcome: { $ref: '#/components/schemas/AutomationConfig' }
        backInStock: { $ref: '#/components/schemas/AutomationConfig' }
      additionalProperties: false
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        details: {}